+++ AppModels/src/FrameResources.cpp
@@ -2,9 +2,9 @@
 
 FrameResources::FrameResources(
 	ComPtr<ID3D12Device> device, 
-	UINT numPassConstants, 
-	UINT numObjectConstants,
-	UINT numMaterialConstants) 
+	size_t numPassConstants,
+	size_t numObjectConstants,
+	size_t numMaterialConstants)
 {
 	m_PassConstantsBuffer = std::make_unique<UploadBuffer<PassConstants>>(device, numPassConstants, true);
 	m_ObjectsConstantsBuffer = std::make_unique<UploadBuffer<ObjectConstants>>(device, numObjectConstants, true);
